requirements.txt
Flask==2.1.2
Flask-Login==0.5.0
Werkzeug==2.1.2
Jinja2==3.1.2
numpy==1.26.4
pillow==10.3.0
torch==2.2.0
torchvision==0.17.0
gunicorn==21.2.0
gdown==5.1.0